fix(redirections): investigate onboarding not showing
