+++ src/stories/Home.stories.tsx
@@ -71,7 +71,5 @@ export default meta;
 type Story = StoryObj<typeof Home>;
 
 export const Default: Story = {
-  args: {
-    setCurrentTab: (tabId: string) => console.log(`Switching to tab: ${tabId}`),
-  },
+  args: {},
 };
